backend/requirements.txt
# Core dependencies for NBA data processing
pandas>=2.0.0
psycopg2-binary>=2.9.0
kagglehub>=0.2.0
python-dotenv>=1.0.0

# Optional: For data analysis and visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
